# GestorBarberia.Api/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GestorBarberia.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Cita/GetCitasById/{clienteId}": {
      "get": {
        "tags": [
          "Cita"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clienteId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cita/GetCitasByCliente/{clienteId}": {
      "get": {
        "tags": [
          "Cita"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clienteId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cita/GetCitasByBarbero/{barberoId}": {
      "get": {
        "tags": [
          "Cita"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "barberoId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cita/Save": {
      "post": {
        "tags": [
          "Cita"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CitaAddDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cita/Update": {
      "put": {
        "tags": [
          "Cita"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CitaUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cita/UpdateEstado": {
      "put": {
        "tags": [
          "Cita"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CitaUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cita/Delete": {
      "delete": {
        "tags": [
          "Cita"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CitaRemoveDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cliente/GetCliente/{name}/{password}": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Cliente/ClienteById/{id}": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cliente/ClienteByBarberiaId/{id}": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cliente/Save": {
      "post": {
        "tags": [
          "Cliente"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ClienteAddDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cliente/Update": {
      "put": {
        "tags": [
          "Cliente"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ClienteUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cliente/Delete": {
      "delete": {
        "tags": [
          "Cliente"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ClienteRemoveDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cliente/DeleteByAdmin/{clienteId}/{adminId}": {
      "delete": {
        "tags": [
          "Cliente"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clienteId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "adminId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Comentario/GetComentariosBarbero/{BarberoId}": {
      "get": {
        "tags": [
          "Comentario"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "BarberoId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Comentario/GetComentariosCliente/{ClienteId}": {
      "get": {
        "tags": [
          "Comentario"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "ClienteId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Comentario/Save": {
      "post": {
        "tags": [
          "Comentario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ComentarioaddDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Comentario/Update": {
      "put": {
        "tags": [
          "Comentario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ComentarioUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Comentario/Delete": {
      "delete": {
        "tags": [
          "Comentario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ComentarioRemoveDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CitaAddDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "citaId": {
          "format": "int32",
          "type": "integer"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "hora": {
          "$ref": "#/definitions/TimeSpan"
        },
        "barberoId": {
          "format": "int32",
          "type": "integer"
        },
        "barberiaId": {
          "format": "int32",
          "type": "integer"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "estiloId": {
          "format": "int32",
          "type": "integer"
        },
        "estado": {
          "type": "string"
        }
      }
    },
    "CitaRemoveDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "citaId": {
          "format": "int32",
          "type": "integer"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "hora": {
          "$ref": "#/definitions/TimeSpan"
        },
        "barberoId": {
          "format": "int32",
          "type": "integer"
        },
        "barberiaId": {
          "format": "int32",
          "type": "integer"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "estiloId": {
          "format": "int32",
          "type": "integer"
        },
        "estado": {
          "type": "string"
        }
      }
    },
    "CitaUpdateDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "citaId": {
          "format": "int32",
          "type": "integer"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "hora": {
          "$ref": "#/definitions/TimeSpan"
        },
        "barberoId": {
          "format": "int32",
          "type": "integer"
        },
        "barberiaId": {
          "format": "int32",
          "type": "integer"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "estiloId": {
          "format": "int32",
          "type": "integer"
        },
        "estado": {
          "type": "string"
        }
      }
    },
    "ClienteAddDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "barberiaId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "imgcliente": {
          "type": "string"
        }
      }
    },
    "ClienteRemoveDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "barberiaId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "imgcliente": {
          "type": "string"
        }
      }
    },
    "ClienteUpdateDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "barberiaId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "imgcliente": {
          "type": "string"
        }
      }
    },
    "ComentarioRemoveDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "idComentarios": {
          "format": "int32",
          "type": "integer"
        },
        "idCita": {
          "format": "int32",
          "type": "integer"
        },
        "idCliente": {
          "format": "int32",
          "type": "integer"
        },
        "idCorte": {
          "format": "int32",
          "type": "integer"
        },
        "idBarbero": {
          "format": "int32",
          "type": "integer"
        },
        "calificacion": {
          "format": "int32",
          "type": "integer"
        },
        "comentario": {
          "type": "string"
        }
      }
    },
    "ComentarioUpdateDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "idComentarios": {
          "format": "int32",
          "type": "integer"
        },
        "idCita": {
          "format": "int32",
          "type": "integer"
        },
        "idCliente": {
          "format": "int32",
          "type": "integer"
        },
        "idCorte": {
          "format": "int32",
          "type": "integer"
        },
        "idBarbero": {
          "format": "int32",
          "type": "integer"
        },
        "calificacion": {
          "format": "int32",
          "type": "integer"
        },
        "comentario": {
          "type": "string"
        }
      }
    },
    "ComentarioaddDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "idComentarios": {
          "format": "int32",
          "type": "integer"
        },
        "idCita": {
          "format": "int32",
          "type": "integer"
        },
        "idCliente": {
          "format": "int32",
          "type": "integer"
        },
        "idCorte": {
          "format": "int32",
          "type": "integer"
        },
        "idBarbero": {
          "format": "int32",
          "type": "integer"
        },
        "calificacion": {
          "format": "int32",
          "type": "integer"
        },
        "comentario": {
          "type": "string"
        }
      }
    },
...
    "TimeSpan": {
      "type": "object",
      "properties": {
        "ticks": {
          "format": "int64",
          "type": "integer"
        },
        "days": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "hours": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "milliseconds": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "microseconds": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "nanoseconds": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "minutes": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "seconds": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "totalDays": {
          "format": "double",
          "type": "number",
          "readOnly": true
        },
        "totalHours": {
          "format": "double",
          "type": "number",
          "readOnly": true
        },
        "totalMilliseconds": {
          "format": "double",
          "type": "number",
          "readOnly": true
        },
        "totalMicroseconds": {
          "format": "double",
          "type": "number",
          "readOnly": true
        },
        "totalNanoseconds": {
          "format": "double",
          "type": "number",
          "readOnly": true
        },
        "totalMinutes": {
          "format": "double",
          "type": "number",
          "readOnly": true
        },
        "totalSeconds": {
          "format": "double",
          "type": "number",
          "readOnly": true
        }
      }
    }
  }
}