# GestorBarberia.Application/Dtos/CitaDto/CitaAddDto.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace GestorBarberia.Application.Dtos.CitaDto
{
    public class CitaAddDto : CitaDto
    {
    }
}