# GestorBarberia.Application/Dtos/ClienteDto/ClienteAddDto.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace GestorBarberia.Application.Dtos.ClienteDto
{
    public class ClienteAddDto : ClienteDto
    {

    }
}