# GestorBarberia.Application/Dtos/ComentarioDto/ComentarioUpdateDto.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace GestorBarberia.Application.Dtos.ComentarioDto
{
    public class ComentarioUpdateDto : ComentarioDto
    {

    }
}